Change volumePerSecond to float

+++ ESP32_WebServer/src/config/config.h
@@ -31,7 +31,7 @@ const unsigned long BLOCK_DURATION_MS = 60000;
 struct PumpSettings {
     uint16_t normalCycleSeconds = 15;
     uint16_t extendedCycleSeconds = 30;
-    uint16_t volumePerSecond = 1;
+    float volumePerSecond = 1.0;
     bool autoModeEnabled = true;
 };
